add text confidentialities / about / terms
