+++ app/Resources/views/politic/about.html.twig
@@ -4,52 +4,31 @@
 
     <div class="container">
         <div class="row">
-            <h1 class="entry-title">Politique de Confidentialité</h1>
+            <h1>A propos</h1>
         </div>
 
         <div class="row">
             <div class="col s12">
+                <h2><strong>Salut !</strong></h2>
 
-                <ul>
-                    <li>Le site cinemove.com est un blog de la société Cybix-Technology, Micro-entreprise au capital de 2000 euros.
-                        <ul>
-                            <li>Adresse: 31b rue Grande, 77670 Saint Mammes</li>
-                            <li>Numéro de Siret: 795 033 901 00039</li>
-                            <li>Non soumis à la TVA</li>
-                            <li>Directeur de la publication: Julien GRIMA</li>
-                            <li>Contact: <a class="fontref" href="mailto:qq1@qqpart.com">cinemove@gmail.com</a></li>
-                            <li>Le site est hébergé par la société O2 SWITCH.</li>
-                            <li></li>
-                        </ul>
-                    </li>
-                </ul>
-                <ul>
-                    <li>Conformément à la loi « informatique et libertés » du 6 janvier 1978 modifiée en 2004, vous bénéficiez d’un droit d’accès et de rectification aux informations qui vous concernent, que vous pouvez exercer en vous adressant à l’administrateur du site, via email en utilisant <a class="fontref" href="mailto:qq1@qqpart.com">cinemove@gmail.com</a>la page de contact.</li>
-                </ul>
-                <p>&nbsp;</p>
-                <ul>
-                    <li>Votre email est collecté pour vous faire parvenir nos nouveaux articles, nouvelles vidéos et nos offres commerciales, dans le strict respect de la réglementation européenne sur la collecte des données. Vos données ne seront jamais revendues à des tiers.</li>
-                </ul>
-                <p>&nbsp;</p>
-                <ul>
-                    <li>Tous les emails envoyés contiennent un lien de désinscription qui permet de ne plus recevoir de message. Les emails collectés sont stockés par <a href="https://www.o2switch.fr/">O2switch</a></li>
-                </ul>
-                <p>&nbsp;</p>
-                <ul>
-                    <li>Cinémove utilise des cookies pour améliorer votre navigation et l’accès au contenu du site. Vous pouvez paramétrer vos cookies en vous référant aux pages support de votre navigateur internet.</li>
-                </ul>
-                <p>&nbsp;</p>
-                <ul>
-                    <li>Vous pouvez également vous opposer au traitement des données vous concernant. Adressez vous à l’administrateur du site,Julien GRIMA, en utilisant <a class="fontref" href="mailto:qq1@qqpart.com">cinemove@gmail.com</a>la page de contact.</li>
-                </ul>
-                <p>&nbsp;</p>
-                <ul>
-                    <li>Vos données personnelles sont conservées de manière confidentielles et ne sont ni cédées ni vendues à des tiers.</li>
-                </ul>
+                <p>Le but de ce site est de vous aider à connaître les sortie de films passées et futures.</p>
+                <p>Je vais vous montrer comment faire.</p>
+                <p>Types, Acteurs, auteurs, synopsis : vous êtes au bon endroit</p>
+
+                <p>Je me suis demandé pourquoi un visiteur voudrait lire cette page… Hmmm, laissez-moi deviner, je vais sortir ma boule de cristal.</p>
+                <p>Vous êtes un vrai fan de films. J'ai bon? Mais cvous n'êtes pas la bibliothèque que vous espériez devenir..</p>
+                <p>Alors vous avez continué à publier du bon contenu, pour vraiment aider vos lecteurs, pour vraiment développer votre audience. Article après article.</p>
+                <p>Dur travail, mais vous sentez que votre cerveau est sur le point d'exploser, non?</p>
+
+                <p><strong>CinéMove est là pour toi.</strong></p>
+
+                <p>La qualité sur Internet est variable. Je dis variable pour être poli. Contenu médiocre. Il y a probablement des fans moins bon que vous, mais qui ont plus de succès.</p>
+                <p>Monde cruel.</p>
+
+                <p>Lance toi et partage tes connaissances inspirées en visitant les pages de CinéMove</p>
 
             </div>
         </div>
-
 
     </div>
