+++ src/AppBundle/Resources/config/routing/about.yml
@@ -5,8 +5,8 @@
 ##############################################
 
 politic:
-    path:     /confidentialities
-    defaults: { _controller: "AppBundle:Default:politic" }
+    path:     /confidentialities/
+    defaults: { _controller: "AppBundle:About:politic" }
 
 ##############################################
 #                                            #
@@ -15,8 +15,8 @@ politic:
 ##############################################
 
 terms:
-    path:     /terms_conditions
-    defaults: { _controller: "AppBundle:Default:terms" }
+    path:     /terms/
+    defaults: { _controller: "AppBundle:About:terms" }
 
 ##############################################
 #                                            #
@@ -25,5 +25,5 @@ terms:
 ##############################################
 
 about:
-    path:     /about
-    defaults: { _controller: "AppBundle:Default:about" }
+    path:     /about/
+    defaults: { _controller: "AppBundle:About:about" }
